requirements.txt
django==3.1.7
django-rest-framework
flake8
requests